+++ client/web/src/cody/components/ChatUI/ChatUi.tsx
@@ -34,6 +34,8 @@ import { GettingStarted } from '../GettingStarted'
 import { ScopeSelector } from '../ScopeSelector'
 import type { ScopeSelectorProps } from '../ScopeSelector/ScopeSelector'
 
+import { useCodyIgnore } from './useCodyIgnore'
+
 import styles from './ChatUi.module.scss'
 
 export const SCROLL_THRESHOLD = 100
@@ -82,6 +84,8 @@ export const ChatUI: React.FC<IChatUIProps> = ({ codyChatStore, isCodyChatPage, 
     const onSubmit = useCallback((text: string) => submitMessage(text), [submitMessage])
     const onEdit = useCallback((text: string) => editMessage(text), [editMessage])
 
+    const { ignores: checkCodyIgnore } = useCodyIgnore()
+
     const scopeSelectorProps: ScopeSelectorProps = useMemo(
         () => ({
             scope,
@@ -92,6 +96,7 @@ export const ChatUI: React.FC<IChatUIProps> = ({ codyChatStore, isCodyChatPage, 
             transcriptHistory,
             className: 'mt-2',
             authenticatedUser,
+            checkCodyIgnore,
         }),
         [
             scope,
@@ -101,6 +106,7 @@ export const ChatUI: React.FC<IChatUIProps> = ({ codyChatStore, isCodyChatPage, 
             logTranscriptEvent,
             transcriptHistory,
             authenticatedUser,
+            checkCodyIgnore,
         ]
     )
