+++ client/web/src/cody/sidebar/Provider.tsx
@@ -1,12 +1,8 @@
 import React, { useContext, useState, useCallback, useMemo } from 'react'
 
-import ignore from 'ignore'
-
-import { useQuery, gql } from '@sourcegraph/http-client'
 import type { AuthenticatedUser } from '@sourcegraph/shared/src/auth'
 import { useTemporarySetting } from '@sourcegraph/shared/src/settings/temporary'
 
-import { parseBrowserRepoURL } from '../../util/url'
 import { useCodyChat, type CodyChatStore, codyChatStoreMock } from '../useCodyChat'
 
 import { useSidebarSize } from './useSidebarSize'
@@ -17,7 +13,6 @@ interface CodySidebarStore extends CodyChatStore {
     setIsSidebarOpen: (isOpen: boolean) => void
     setFocusProvided: () => void
     setSidebarSize: (size: number) => void
-    ignores: (path: string) => boolean
 }
 
 const CodySidebarContext = React.createContext<CodySidebarStore | null>({
@@ -27,7 +22,6 @@ const CodySidebarContext = React.createContext<CodySidebarStore | null>({
     setSidebarSize: () => {},
     setIsSidebarOpen: () => {},
     setFocusProvided: () => {},
-    ignores: () => false,
 })
 
 interface ICodySidebarStoreProviderProps {
@@ -35,43 +29,7 @@ interface ICodySidebarStoreProviderProps {
     authenticatedUser: AuthenticatedUser | null
 }
 
-const MY_QUERY = gql`
-    query CodyIgnoreContent($repoName: String!, $repoRev: String!, $filePath: String!) {
-        repository(name: $repoName) {
-            commit(rev: $repoRev) {
-                blob(path: $filePath) {
-                    content
-                }
-            }
-        }
-    }
-`
-
-const CODY_IGNORE_PATH = '.cody/ignore'
-const useCodyIgnore = (): { ignores: (path: string) => boolean } => {
-    const { repoName, revision } = parseBrowserRepoURL(location.pathname + location.search + location.hash)
-    const { data } = useQuery<any, any>(MY_QUERY, {
-        variables: { repoName, repoRev: revision, filePath: CODY_IGNORE_PATH },
-    })
-    const ignoreManager = useMemo(() => (data ? ignore().add(data.repository.commit.blob.content || '') : null), [data])
-
-    // TODO: remove
-    window.ignoreManager = ignoreManager
-
-    const ignores = useCallback(
-        (path: string): boolean => {
-            if (ignoreManager) {
-                return ignoreManager.ignores(path)
-            }
-            return false
-        },
-        [ignoreManager]
-    )
-    return { ignores }
-}
-
 export const CodySidebarStoreProvider: React.FC<ICodySidebarStoreProviderProps> = ({ authenticatedUser, children }) => {
-    const { ignores } = useCodyIgnore()
     const [isSidebarOpen, setIsSidebarOpenState] = useTemporarySetting('cody.showSidebar', false)
     const [inputNeedsFocus, setInputNeedsFocus] = useState(false)
     const { setSidebarSize } = useSidebarSize()
@@ -100,9 +58,8 @@ export const CodySidebarStoreProvider: React.FC<ICodySidebarStoreProviderProps> 
             setIsSidebarOpen,
             setFocusProvided,
             setSidebarSize,
-            ignores,
         }),
-        [codyChatStore, isSidebarOpen, setIsSidebarOpen, setFocusProvided, setSidebarSize, inputNeedsFocus, ignores]
+        [codyChatStore, isSidebarOpen, setIsSidebarOpen, setFocusProvided, setSidebarSize, inputNeedsFocus]
     )
 
     // dirty fix because CodyRecipesWidget is rendered inside a different React DOM tree.
